memoize hook variables as needed
