+++ context/app/static/js/components/publications/hooks.ts
@@ -1,7 +1,9 @@
+import { useMemo, useState } from 'react';
 import { ContributorAPIResponse } from 'js/components/detailPage/ContributorsTable/utils';
 import { useSearchHits } from 'js/hooks/useSearchData';
 import { useDownloadTable } from 'js/helpers/download';
 import { usePublicationsSearchState } from 'js/components/publications/PublicationsSearchContext';
+import { useEventCallback } from '@mui/material/utils';
 
 export interface Publication {
   uuid: string;
@@ -39,25 +41,47 @@ function usePublications() {
   const { publications = [], isLoading } = usePublicationHits();
   const search = usePublicationsSearchState();
 
+  const filteredPublications = useMemo(() => {
+    return publications
+      .filter((item) => item.title.toLowerCase().includes(search.toLowerCase()))
+      .sort((a, b) => new Date(b.publication_date).getTime() - new Date(a.publication_date).getTime());
+  }, [publications, search]);
+
+  const publicationsByStatus = useMemo(() => {
+    const preprints = filteredPublications.filter((pub) => !pub.publication_status);
+    const reviewed = filteredPublications.filter((pub) => pub.publication_status);
+
+    return [
+      { status: 'Peer-Reviewed', relevantPublications: reviewed },
+      { status: 'Preprint', relevantPublications: preprints },
+    ];
+  }, [filteredPublications]);
+
   const downloadTable = useDownloadTable({
     fileName: 'publications.tsv',
     columnNames: ['Title', 'Contributors', 'Publication Venue', 'Published Date'],
-    rows: publications.map(({ title, contributors, publication_venue, publication_date }) => {
-      const contributorsList = contributors
-        ? contributors.map(({ first_name, last_name }) => `${first_name} ${last_name}`).join(', ')
-        : '';
-      return [title, contributorsList, publication_venue, publication_date];
-    }),
+    rows: useMemo(() => {
+      return publications.map(({ title, contributors, publication_venue, publication_date }) => {
+        const contributorsList = contributors
+          ? contributors.map(({ first_name, last_name }) => `${first_name} ${last_name}`).join(', ')
+          : '';
+        return [title, contributorsList, publication_venue, publication_date];
+      });
+    }, [publications]),
   });
 
-  const filteredPublications = publications
-    .filter((item) => item.title.toLowerCase().includes(search.toLowerCase()))
-    .sort((a, b) => new Date(b.publication_date).getTime() - new Date(a.publication_date).getTime());
+  const [openTabIndex, setOpenTabIndex] = useState(0);
+
+  const handleTabChange = useEventCallback((_: Event, newIndex: number) => {
+    setOpenTabIndex(newIndex);
+  });
 
   return {
-    publications,
-    filteredPublications,
     isLoading,
+    filteredPublications,
+    publicationsByStatus,
+    openTabIndex,
+    handleTabChange,
     downloadTable,
   };
 }
